New project structure (blueprint), database classes, flask cmd create_tables
