+++ .env
@@ -1,0 +1,1 @@
+SECRET_KEY = '248fb9a5bdffa13c0bc136504ebf75c2'
