+++ .flaskenv
@@ -1,0 +1,3 @@
+FLASK_APP=main_app
+FLASK_DEBUG=True
+FLASK_ENV=Development
