--- app.py
@@ -1,6 +1,0 @@
-from main_app import app
-
-
-if __name__ == '__main__':
-    app.debug=True
-    app.run()
