+++ app.py.old
@@ -1,0 +1,6 @@
+from main_app import app
+
+
+if __name__ == '__main__':
+    app.debug=True
+    app.run()
