--- config.py
@@ -1,2 +1,0 @@
-SECRET_KEY = '248fb9a5bdffa13c0bc136504ebf75c2'
-SQLALCHEMY_DATABASE_URI = 'sqlite:///site.db'
