--- db_create.py
@@ -1,2 +1,0 @@
-from main_app import db
-db.create_all()
