+++ db_create.py.old
@@ -1,0 +1,2 @@
+from main_app import db
+db.create_all()
