+++ main_app/algemeen.py.old
@@ -1,0 +1,7 @@
+from flask import Blueprint, render_template
+
+algemeen_bp = Blueprint("algemeen_bp", __name__)
+
+# @algemeen_bp.route("/")
+# def index():
+    # return render_template("index.html")
