+++ main_app/templates/index.html
@@ -29,33 +29,33 @@
 
                 <input type="radio" class="btn-check" name="btnradio" id="btnradio2" autocomplete="off">
                 <label class="btn btn-outline-primary" for="btnradio2">
-                    <a href=" {{ url_for('login') }} " class="nav-link"> Login  </a>
+                    <a href=" {{ url_for('login_bp.login') }} " class="nav-link"> Login  </a>
                 </label>
 
                 <input type="radio" class="btn-check" name="btnradio" id="btnradio3" autocomplete="off">
                 <label class="btn btn-outline-primary" for="btnradio3">
-                    <a href=" {{ url_for('loginParents') }} " class="nav-link">
+                    <a href=" {{ url_for('login_parents_bp.loginParents') }} " class="nav-link">
                         Parents Login
                     </a>
                 </label>
 
                 <input type="radio" class="btn-check" name="btnradio" id="btnradio4" autocomplete="off">
                 <label class="btn btn-outline-primary" for="btnradio4">
-                    <a href=" {{ url_for('admission') }} " class="nav-link">
+                    <a href=" {{ url_for('admission_bp.admission') }} " class="nav-link">
                         Admission
                     </a>
                 </label>
 
                 <input type="radio" class="btn-check" name="btnradio" id="btnradio5" autocomplete="off">
                 <label class="btn btn-outline-primary" for="btnradio5">
-                    <a href=" {{ url_for('applyParents') }} " class="nav-link">
+                    <a href=" {{ url_for('apply_parents_bp.apply_parents') }} " class="nav-link">
                         Parents Application Form
                     </a>
                 </label>
 
                 <input type="radio" class="btn-check" name="btnradio" id="btnradio6" autocomplete="off">
                 <label class="btn btn-outline-primary" for="btnradio6">
-                    <a href=" {{ url_for('applyTeacher') }} " class="nav-link">
+                    <a href=" {{ url_for('apply_teacher_bp.apply_teacher') }} " class="nav-link">
                         Teacher Application Form
                     </a>
                 </label>
