+++ requirements.txt.old
@@ -18,10 +18,9 @@ itsdangerous==1.1.0
 Jinja2==2.11.2
 Mako==1.1.3
 MarkupSafe==1.1.1
-pkg-resources==0.0.0
+#psycopg2==2.8.6
 pycparser==2.20
 python-dateutil==2.8.1
-python-dotenv==0.15.0
 python-editor==1.0.4
 six==1.15.0
 SQLAlchemy==1.3.22
